split out util to its own file, template out the identifier text.
