+++ demos/api/core_web_vitals/main.js
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-const fetch = require('node-fetch');
+const CrUXApiUtil = require('./cruxApiUtil.js');
 const chalk = require('chalk');
 var argv = require('yargs').options({
   'o': {
@@ -36,15 +36,11 @@ var argv = require('yargs').options({
     type: 'string',
   }}).argv;
 
-// Setup the API Util object.
-const CrUXApiUtil = {};
-
-// Setup the API key.
+// Confirm the API key.
 if (!argv.key) {
   console.log('An API Key is required to use the Chrome UX Report API.');
   return;
 }
-CrUXApiUtil.KEY = argv.key;
 
 // Create request to send to API.
 let request = {}
@@ -64,90 +60,11 @@ if (argv.formFactor) {
   request.formFactor = argv.formFactor;
 }
 
-/**
- * Query the CrUX API for metric records.
- *
- * @param {{origin?: string, url?: string, formFactor?: string,  metrics?: Array<string>}} requestBody
- * return {{record: {key: Object, metrics: Map<string, {histogram: Array<{ start: number, end?: number, density: number}>}>}, urlNormalizationDetails: Object}} */
-CrUXApiUtil.query = async function (requestBody) {
-  const endpointUrl = 'https://chromeuxreport.googleapis.com/v1/records:queryRecord';
-  const resp = await fetch(`${endpointUrl}?key=${CrUXApiUtil.KEY}`, {
-    method: 'POST',
-    body: JSON.stringify(requestBody),
-  });
-
-  const json = await resp.json();
-  if (!resp.ok) {
-    console.log(chalk.red(json.error.message));
-    if (resp.status === 404) {
-      // Extra Not Found flavor text.
-      console.log('Data cannot be found for this URL. ' +
-        'Check that the URL you\'re looking up is correct, ' +
-        'don\'t forget "www" if you need it!');
-    }
-    process.exit(1);
-  }
-  return json;
-};
-
-/**
- * Utility method to transform the response's metric data into an array of usable metric objects
- * Example return value:
-  [
-    {
-      "acronym": "FCP",
-      "name": "first_contentful_paint",
-      "labeledBins": [{
-        "label": "good",
-        "percentage": 43.51,
-        "start": 0,
-        "end": 1000,
-        "density": 0.4351
-      }, // ... (other bins) ]
-      "percentiles": {
-          "p75": 1792
-      },
-    }, // ... (other metrics) ]
-*
-* @return {{acronym: string, name: string, labelsBins: Array<{label: 'good'|'needs improvement'|'poor', percentage: number, start: number, end?: number, density: number}>}}*/
-function labelMetricData(metrics) {
-  const nameToAcronymMap = {
-    first_contentful_paint: 'FCP',
-    largest_contentful_paint: 'LCP',
-    first_input_delay: 'FID',
-    cumulative_layout_shift: 'CLS',
-  };
-
-  return Object.entries(metrics).map(([metricName, metricData]) => {
-    const standardBinLabels = ['good', 'needs improvement', 'poor'];
-    const metricBins = metricData.histogram;
-
-    // We assume there are 3 histogram bins and they're in order of: good => poor
-    const labeledBins = metricBins.map((bin, i) => {
-      // Assign a good/poor label, calculate a percentage, and add retain all existing bin properties
-      return {
-        label: standardBinLabels[i],
-        percentage: bin.density * 100,
-        ...bin,
-      };
-    });
-
-    return {
-      acronym: nameToAcronymMap[metricName],
-      name: metricName,
-      labeledBins,
-      percentiles: {
-          p75: metricData.percentiles.p75,
-      },
-    };
-  });
-}
-
 // Gather the data and display it.
 (async function () {
-  const json = await CrUXApiUtil.query(request);
+  const json = await CrUXApiUtil.query(request, argv.key);
 
-  const labeledMetrics = await labelMetricData(json.record.metrics);
+  const labeledMetrics = await CrUXApiUtil.labelMetricData(json.record.metrics);
 
   // The CWV 75%ile thresholds.
   const nameToCWVMap = {
@@ -157,20 +74,18 @@ function labelMetricData(metrics) {
   };
 
   // Display metric results
-  let resId = '';
-  let identifier = '';
-  if (json.record.key.origin) {
-    resId += 'Origin: ' + json.record.key.origin;
-    identifier = json.record.key.origin;
-  } else {
-    resId += 'Url: ' + json.record.key.url;
+  let identifierType = 'Origin';
+  let identifier = json.record.key.origin;
+  if (json.record.key.url) {
+    identifierType = 'Url';
     identifier = json.record.key.url;
   }
+  let formFactor = 'ALL';
   if (json.record.key.formFactor) {
-    resId += ', FormFactor: ' + json.record.key.formFactor;
+    formFactor = json.record.key.formFactor;
   }
   // Add form factor too...
-  console.log('\n' + resId);
+  console.log('\n' + `${identifierType}: ${identifier}, FormFactor: ${formFactor}`);
   for (const metric of labeledMetrics) {
     if (nameToCWVMap[metric.name]) {
         // Check pass fail
